+++ omniva-woocommerce/js/omniva.js
@@ -25,16 +25,18 @@ jQuery('document').ready(function($){
         });
     });
 
-    $( document ).on( 'omnivalt.checkpostcode, updated_checkout', function() {
+    if (typeof omnivaSettings !== 'undefined') {
+    	$( document ).on( 'omnivalt.checkpostcode, updated_checkout', function() {
         if (omnivaSettings.auto_select == "yes" && omniva_current_terminal === '') {
             current_postcode = select_terminal(current_postcode);
         }
-    });
-    $( document ).on('change', '#shipping_postcode, #billing_postcode',function(){
+    	});
+    	$( document ).on('change', '#shipping_postcode, #billing_postcode',function(){
         if (omnivaSettings.auto_select == "yes" && omniva_current_terminal === '') {
             current_postcode = select_terminal(current_postcode);
         }
-    });  
+    	});
+    }  
     $( document.body ).on( 'updated_wc_div', function(){
         //$('.omnivalt_terminal').omniva();
         $("select.shipping_method, :input[name^=shipping_method]:checked").trigger('change'); //TODO: Need better solution for dropdown update when in cart change country